third_party/blink/web_tests/security/cannot-read-self-from-file.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
testRunner.setAllowFileAccessFromFileURLs(false);

var t = async_test('file: should be a unique-origin protocol for XHR purposes');
window.addEventListener('message', t.step_func(function(evt) {
  assert_equals(evt.data, 'PASS');
  t.done();
}));
</script>
</head>
<body>
<iframe src="resources/cannot-read-self-from-file.html"></iframe>
</body>
</html>
